Moved AccordionPoll to a module folder, moved events out of core
