+++ lib/src/core/command_statistics.dart
@@ -4,7 +4,7 @@ import 'package:weather/src/core/repositories/command_statistics_repository.dart
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/injector/injection.dart';
 import 'package:weather/src/utils/wait_concurrently.dart';
-import 'events/access_events.dart';
+import 'package:weather/src/events/access_events.dart';
 import 'event_bus.dart';
 import 'chat.dart';
