+++ lib/src/modules/accordion_poll/accordion_poll_manager.dart
@@ -1,10 +1,10 @@
-import 'package:weather/src/core/accordion_poll.dart';
 import 'package:weather/src/core/user.dart';
 import 'package:weather/src/core/chat.dart';
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/platform/platform.dart';
-import 'utils.dart';
+import 'accordion_poll.dart';
+import '../utils.dart';
 
 class AccordionPollManager {
   final Platform platform;
